+++ docs/html/search/all_10.js
@@ -1,7 +1,6 @@
 var searchData=
 [
-  ['release_20notes_0',['Release notes',['../index.html#autotoc_md16',1,'']]],
-  ['releasenotes_1',['RELEASENOTES',['../md__2builds_2oei_2dlb__alps__native_2RELEASENOTES.html',1,'']]],
-  ['requirements_2',['&lt;strong&gt;Dolby&lt;/strong&gt; AC-4 content requirements',['../index.html#autotoc_md12',1,'']]],
-  ['return_5fvalues_5ft_3',['return_values_t',['../structreturn__values__t.html',1,'']]]
+  ['native_20c_0',['ALPS Native (C)',['../index.html',1,'']]],
+  ['notes_1',['Release notes',['../index.html#autotoc_md16',1,'']]],
+  ['num_5fentities_5fin_5fgroup_2',['num_entities_in_group',['../structalps__mp4dmx__preselection__t.html#a976dae7d0f42b4072568d834987e9371',1,'alps_mp4dmx_preselection_t']]]
 ];