+++ docs/html/search/all_11.js
@@ -1,13 +1,5 @@
 var searchData=
 [
-  ['scheme_5furi_0',['scheme_uri',['../structalps__mp4dmx__prsl__kind__t.html#a19de8c79b17cf1ae01ae860e25298599',1,'alps_mp4dmx_prsl_kind_t::scheme_uri'],['../structalps__presentation__kind__t.html#a1f504f5cc23bb601b8ee2759fae92570',1,'alps_presentation_kind_t::scheme_uri']]],
-  ['segment_20signaling_1',['ISOBMFF init segment signaling',['../index.html#autotoc_md8',1,'']]],
-  ['segment_5fbuf_2',['segment_buf',['../structalps__mp4dmx__t.html#a0e6a45a506f9dfb07682679685cd97d6',1,'alps_mp4dmx_t']]],
-  ['segment_5fbuf_5fpos_3',['segment_buf_pos',['../structalps__mp4dmx__t.html#a457170a7ae4081f73865334f5239693c',1,'alps_mp4dmx_t']]],
-  ['selection_5fpriority_4',['selection_priority',['../structalps__mp4dmx__preselection__t.html#a93e111db5081352fa6ca9ea673d3bd23',1,'alps_mp4dmx_preselection_t::selection_priority'],['../structalps__presentation__t.html#aa67a883a2d18fcfcba87af618acd46cb',1,'alps_presentation_t::selection_priority']]],
-  ['signaling_5',['signaling',['../index.html#autotoc_md7',1,'&lt;strong&gt;Dolby&lt;/strong&gt; AC-4 presentation signaling'],['../index.html#autotoc_md9',1,'External signaling'],['../index.html#autotoc_md8',1,'ISOBMFF init segment signaling']]],
-  ['strong_20dolby_20strong_20ac_204_20content_20requirements_6',['&lt;strong&gt;Dolby&lt;/strong&gt; AC-4 content requirements',['../index.html#autotoc_md12',1,'']]],
-  ['strong_20dolby_20strong_20ac_204_20presentation_20signaling_7',['&lt;strong&gt;Dolby&lt;/strong&gt; AC-4 presentation signaling',['../index.html#autotoc_md7',1,'']]],
-  ['super_8',['super',['../structbam__binary__sink__t.html#a433ef2869cf81fe212c29b5ff0b0b435',1,'bam_binary_sink_t']]],
-  ['support_20contact_9',['Support / Contact',['../index.html#autotoc_md14',1,'']]]
+  ['of_20contents_0',['Table of Contents',['../index.html#autotoc_md1',1,'']]],
+  ['offset_1',['offset',['../structpresentation__level__t.html#af159b9ea1fd08cf108cc322a3d8bb52c',1,'presentation_level_t']]]
 ];