+++ docs/html/search/all_13.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['utils_2eh_0',['utils.h',['../utils_8h.html',1,'']]]
+  ['quickstart_0',['Quickstart',['../index.html#autotoc_md3',1,'']]]
 ];