+++ docs/html/search/all_14.js
@@ -1,5 +1,6 @@
 var searchData=
 [
-  ['value_0',['value',['../structpresentation__level__t.html#a1a278c87918aecfb5b8ba77911469625',1,'presentation_level_t::value'],['../structalps__mp4dmx__prsl__kind__t.html#a3f38e4b6bbefc6a212ad7c16d9457abf',1,'alps_mp4dmx_prsl_kind_t::value'],['../structalps__presentation__kind__t.html#acd2df7e16d2f951ae776a784121daef8',1,'alps_presentation_kind_t::value']]],
-  ['version_1',['version',['../structalps__mp4dmx__full__box__t.html#ae56a4fa06c789f619c8b5c435411c33b',1,'alps_mp4dmx_full_box_t']]]
+  ['release_20notes_0',['Release notes',['../index.html#autotoc_md16',1,'']]],
+  ['requirements_1',['&lt;strong&gt;Dolby&lt;/strong&gt; AC-4 content requirements',['../index.html#autotoc_md12',1,'']]],
+  ['return_5fvalues_5ft_2',['return_values_t',['../structreturn__values__t.html',1,'']]]
 ];