+++ docs/html/search/all_4.js
@@ -1,9 +1,5 @@
 var searchData=
 [
-  ['dialog_5fgain_0',['dialog_gain',['../structalps__mp4dmx__preselection__t.html#a6bcba8996af9a99ccbe72f36387bec46',1,'alps_mp4dmx_preselection_t::dialog_gain'],['../structalps__presentation__t.html#a9cdafa5ad65891cf9af9e5fe110ee118',1,'alps_presentation_t::dialog_gain']]],
-  ['dialog_5fgain_5fpresent_1',['dialog_gain_present',['../structalps__mp4dmx__preselection__t.html#acd82726dfa1ae75b6e9a48ccd69925ac',1,'alps_mp4dmx_preselection_t::dialog_gain_present'],['../structalps__presentation__t.html#a9a47a7881381c33f240208f8d2494f94',1,'alps_presentation_t::dialog_gain_present']]],
-  ['dlb_5falps_5fnative_2ec_2',['dlb_alps_native.c',['../dlb__alps__native_8c.html',1,'']]],
-  ['dlb_5falps_5fnative_2eh_3',['dlb_alps_native.h',['../dlb__alps__native_8h.html',1,'']]],
-  ['dolby_20strong_20ac_204_20content_20requirements_4',['&lt;strong&gt;Dolby&lt;/strong&gt; AC-4 content requirements',['../index.html#autotoc_md12',1,'']]],
-  ['dolby_20strong_20ac_204_20presentation_20signaling_5',['&lt;strong&gt;Dolby&lt;/strong&gt; AC-4 presentation signaling',['../index.html#autotoc_md7',1,'']]]
+  ['4_20content_20requirements_0',['&lt;strong&gt;Dolby&lt;/strong&gt; AC-4 content requirements',['../index.html#autotoc_md12',1,'']]],
+  ['4_20presentation_20signaling_1',['&lt;strong&gt;Dolby&lt;/strong&gt; AC-4 presentation signaling',['../index.html#autotoc_md7',1,'']]]
 ];