+++ docs/html/search/all_8.js
@@ -1,7 +1,9 @@
 var searchData=
 [
-  ['init_20segment_20signaling_0',['ISOBMFF init segment signaling',['../index.html#autotoc_md8',1,'']]],
-  ['installation_1',['Installation',['../index.html#autotoc_md2',1,'']]],
-  ['is_5fgroup_5flabel_2',['is_group_label',['../structalps__mp4dmx__prsl__labl__t.html#a1a5445647fc6eb5acf74457dfded0308',1,'alps_mp4dmx_prsl_labl_t::is_group_label'],['../structalps__presentation__label__t.html#ab2687acd639afa9c4a2e3ef1770243ae',1,'alps_presentation_label_t::is_group_label']]],
-  ['isobmff_20init_20segment_20signaling_3',['ISOBMFF init segment signaling',['../index.html#autotoc_md8',1,'']]]
+  ['dialog_5fgain_0',['dialog_gain',['../structalps__mp4dmx__preselection__t.html#a6bcba8996af9a99ccbe72f36387bec46',1,'alps_mp4dmx_preselection_t::dialog_gain'],['../structalps__presentation__t.html#a9cdafa5ad65891cf9af9e5fe110ee118',1,'alps_presentation_t::dialog_gain']]],
+  ['dialog_5fgain_5fpresent_1',['dialog_gain_present',['../structalps__mp4dmx__preselection__t.html#acd82726dfa1ae75b6e9a48ccd69925ac',1,'alps_mp4dmx_preselection_t::dialog_gain_present'],['../structalps__presentation__t.html#a9a47a7881381c33f240208f8d2494f94',1,'alps_presentation_t::dialog_gain_present']]],
+  ['dlb_5falps_5fnative_2ec_2',['dlb_alps_native.c',['../dlb__alps__native_8c.html',1,'']]],
+  ['dlb_5falps_5fnative_2eh_3',['dlb_alps_native.h',['../dlb__alps__native_8h.html',1,'']]],
+  ['dolby_20strong_20ac_204_20content_20requirements_4',['&lt;strong&gt;Dolby&lt;/strong&gt; AC-4 content requirements',['../index.html#autotoc_md12',1,'']]],
+  ['dolby_20strong_20ac_204_20presentation_20signaling_5',['&lt;strong&gt;Dolby&lt;/strong&gt; AC-4 presentation signaling',['../index.html#autotoc_md7',1,'']]]
 ];