+++ docs/html/search/all_9.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['kinds_0',['kinds',['../structalps__mp4dmx__preselection__t.html#a82f3562181fbeb92c1a68a3e2aad10f3',1,'alps_mp4dmx_preselection_t::kinds'],['../structalps__presentation__t.html#abe46f42a43cad6040358637ffa1398da',1,'alps_presentation_t::kinds']]],
-  ['kinds_5fcount_1',['kinds_count',['../structalps__mp4dmx__preselection__t.html#a01ddf77fd7ed3ab3427a8044a32039f5',1,'alps_mp4dmx_preselection_t::kinds_count'],['../structalps__presentation__t.html#afce7feb6f18e5ea10a1837fcf23dd221',1,'alps_presentation_t::kinds_count']]],
-  ['known_20limitations_2',['Known limitations',['../index.html#autotoc_md10',1,'']]]
+  ['entity_5fids_0',['entity_ids',['../structalps__mp4dmx__preselection__t.html#a5e463f4b0fd217caa2187c85de3a8df8',1,'alps_mp4dmx_preselection_t']]],
+  ['extended_5flanguage_1',['extended_language',['../structalps__mp4dmx__preselection__t.html#aca110d2177b95120b04b568cefeffd04',1,'alps_mp4dmx_preselection_t::extended_language'],['../structalps__presentation__t.html#a2689261f4b7030e0eb3ee27b01d15869',1,'alps_presentation_t::extended_language']]],
+  ['external_20signaling_2',['External signaling',['../index.html#autotoc_md9',1,'']]]
 ];