+++ docs/html/search/all_e.js
@@ -1,14 +1,12 @@
 var searchData=
 [
-  ['party_0',['License &amp; third-party',['../index.html#autotoc_md15',1,'']]],
-  ['preselection_5ftag_1',['preselection_tag',['../structalps__mp4dmx__preselection__t.html#a015776983479b6b523201529ebee3ff0',1,'alps_mp4dmx_preselection_t']]],
-  ['preselections_2',['preselections',['../structalps__mp4dmx__t.html#aa8b6f514515d8e738ccda96b053b2014',1,'alps_mp4dmx_t::preselections'],['../structalps__mp4dmx__grpl__t.html#ae5e09a27bd6b92c89fddb65a2bbb6348',1,'alps_mp4dmx_grpl_t::preselections']]],
-  ['preselections_5fcount_3',['preselections_count',['../structalps__mp4dmx__t.html#af323d81cd9cefe85074068fc692c4348',1,'alps_mp4dmx_t::preselections_count'],['../structalps__mp4dmx__grpl__t.html#a7e30c85719dcb1a7e170d33a297fde0d',1,'alps_mp4dmx_grpl_t::preselections_count']]],
-  ['presentation_20signaling_4',['&lt;strong&gt;Dolby&lt;/strong&gt; AC-4 presentation signaling',['../index.html#autotoc_md7',1,'']]],
-  ['presentation_5flevel_5',['presentation_level',['../bam__binary__sink_8c.html#a054ed908f1c03dbcb8d9e9d7f6594cb5',1,'bam_binary_sink.c']]],
-  ['presentation_5flevel_5ft_6',['presentation_level_t',['../structpresentation__level__t.html',1,'']]],
-  ['presentations_7',['presentations',['../structbam__binary__sink__t.html#abb9e815ca82f0f9d1c3c5ca180a61358',1,'bam_binary_sink_t::presentations'],['../structalps__ctx__t.html#a4565bbd8a2dc092e77b4055cc61e706a',1,'alps_ctx_t::presentations']]],
-  ['presentations_5fcb_8',['presentations_cb',['../structalps__ctx__t.html#a05b2a0103c0b0d7419c38e695970d0af',1,'alps_ctx_t']]],
-  ['presentations_5fcb_5fctx_9',['presentations_cb_ctx',['../structalps__ctx__t.html#a374a6054e690600805de87fb4cc83b36',1,'alps_ctx_t']]],
-  ['presentations_5fcount_10',['presentations_count',['../structbam__binary__sink__t.html#a537833f24759b2dd40f0e7ec100e49d4',1,'bam_binary_sink_t::presentations_count'],['../structalps__ctx__t.html#a266059cd09598f03bd88e5d300e4d1ae',1,'alps_ctx_t::presentations_count']]]
+  ['label_0',['label',['../structalps__mp4dmx__prsl__labl__t.html#abf5bc15f90712f2790dc55e130edc045',1,'alps_mp4dmx_prsl_labl_t::label'],['../structalps__presentation__label__t.html#a9fe3be0cc49be2331140f1574a98314b',1,'alps_presentation_label_t::label']]],
+  ['label_5fid_1',['label_id',['../structalps__mp4dmx__prsl__labl__t.html#a3f6e9c04fe4339757aa2cf13ceaf0cef',1,'alps_mp4dmx_prsl_labl_t::label_id'],['../structalps__presentation__label__t.html#a157359c4dde91b9600a2cf03361c7077',1,'alps_presentation_label_t::label_id']]],
+  ['labels_2',['labels',['../structalps__mp4dmx__preselection__t.html#a4768786ee888521ea0538ac2d32c11e0',1,'alps_mp4dmx_preselection_t::labels'],['../structalps__presentation__t.html#a93982f8b0c9238024febf8892855ade8',1,'alps_presentation_t::labels']]],
+  ['labels_5fcount_3',['labels_count',['../structalps__mp4dmx__preselection__t.html#ad08f6b04eee0c6c021dfa2438357aa11',1,'alps_mp4dmx_preselection_t::labels_count'],['../structalps__presentation__t.html#af185137e91b5af2a34090981d49126d7',1,'alps_presentation_t::labels_count']]],
+  ['language_4',['language',['../structalps__mp4dmx__prsl__labl__t.html#ad63235ecea932302a273f1b1117ddf59',1,'alps_mp4dmx_prsl_labl_t::language'],['../structalps__presentation__label__t.html#a7775c5fcb389a7a35ed3afeb87d497e8',1,'alps_presentation_label_t::language']]],
+  ['levels_5',['levels',['../structbam__binary__sink__t.html#ad1a9cf3c0c687e7a778db4feb6bc65cc',1,'bam_binary_sink_t']]],
+  ['library_6',['Library',['../index.html#autotoc_md4',1,'']]],
+  ['license_20third_20party_7',['License &amp; third-party',['../index.html#autotoc_md15',1,'']]],
+  ['limitations_8',['Known limitations',['../index.html#autotoc_md10',1,'']]]
 ];