+++ docs/html/search/pages_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['alps_20native_20c_0',['ALPS Native (C)',['../index.html',1,'']]]
+  ['0_200_0',['3.0.0',['../md__2builds_2oei_2dlb__alps__native_2RELEASENOTES.html',1,'']]]
 ];