+++ docs/html/search/pages_2.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['native_20c_0',['ALPS Native (C)',['../index.html',1,'']]]
+  ['alps_20native_20c_0',['ALPS Native (C)',['../index.html',1,'']]]
 ];