+++ docs/html/search/pages_3.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['releasenotes_0',['RELEASENOTES',['../md__2builds_2oei_2dlb__alps__native_2RELEASENOTES.html',1,'']]]
+  ['c_0',['ALPS Native (C)',['../index.html',1,'']]]
 ];