+++ docs/html/search/searchdata.js
@@ -1,6 +1,6 @@
 var indexSectionsWithContent =
 {
-  0: "4abcdefgiklmnopqrstuv",
+  0: "01234abcdefgiklmnopqrstuv",
   1: "abprt",
   2: "abdtu",
   3: "ab",
@@ -9,7 +9,7 @@ var indexSectionsWithContent =
   6: "a",
   7: "a",
   8: "abc",
-  9: "acnr"
+  9: "03acn"
 };
 
 var indexSectionNames =